+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+## 1.0.3
+
+- UIComponent.parentUIComponent
+- UIComponent.onChildRendered()
+- Navigation.parameterAsInt/parameterAsNum/parameterAsBool
+- Navigation.parameterAsStringList/parameterAsIntList/parameterAsNumList/parameterAsBoolList
+- UINavigator._encodeRouteParameters(): Comma ',' won't be encoded as %2C
+- swiss_knife: ^2.3.5
+
 ## 1.0.2
 
 - of prefix 'ui-' for css.